构建：完成 Windows x64 C++23 编译适配
统一仓库受跟踪文本为 UTF-8，增加编码校验脚本与编辑器规则，并为 MSVC 强制启用 /utf-8，从源头避免第三方头文件、配置和源码因非 UTF-8 字节导致编译失败。

处理 Windows 头文件宏与安全接口差异：隔离 min/max、GetObject 等宏污染，修复 CTP 加载器依赖与链接关系，安全封装固定缓冲区复制、错误文本、模块名称、崩溃转储路径、调用栈回调和跨平台时间解析。

按 WonderTrader 上下游数据契约修正隐式收窄和越界风险：统一数组、算法、共享内存、历史数据与 K 线索引宽度；保留 HFT/UFT 数量语义；校验 HuaX 委托数量、OES 固定字段及 ATP 价格、序号、回报索引和分页精度。

完善 Windows 构建系统：支持 Visual Studio 多配置输出和复制目标，按 PE 机器类型筛选第三方运行库，部署系统与 vcpkg 依赖，修复安装后的运行时搜索路径，并保证 CTP/CTPOpt 等插件在 NOMINMAX 下可独立构建。

补充 Windows 与跨平台回归测试，覆盖 UTF-8、固定字段边界、共享内存布局、HFT/UFT 订单、ATP/HuaX 数值转换、DLL 加载、运行库架构、安装冒烟和多配置复制；最终在 Windows amd64 Release C++23 /WX 下完成全量构建及 35 项 CTest。

+++ src/WtExeFact/WtStockMinImpactExeUnit.h
@@ -177,7 +177,7 @@ private:
 	uint64_t _now;
 
 public:
-	inline int round_hands(double hands, double min_hands)
+	inline double round_hands(double hands, double min_hands)
 	{
 		return (int)((hands + min_hands / 2) / min_hands) * min_hands;
 	}
